[Chore] #64 - HomeViewController 코드 분리
